Upgrade/cleanup C# sample
