--- C#/TrueProfile.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrueProfile", "TrueProfile\TrueProfile.csproj", "{C698E55E-9A99-499D-8C77-165B52436DE3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C698E55E-9A99-499D-8C77-165B52436DE3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C698E55E-9A99-499D-8C77-165B52436DE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C698E55E-9A99-499D-8C77-165B52436DE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C698E55E-9A99-499D-8C77-165B52436DE3}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
