--- C#/TrueProfile.userprefs
@@ -1,14 +1,0 @@
-﻿<Properties StartupConfiguration="{C698E55E-9A99-499D-8C77-165B52436DE3}|Default">
-  <MonoDevelop.Ide.Workbench ActiveDocument="TrueProfile/Program.cs">
-    <Files>
-      <File FileName="TrueProfile/Program.cs" Line="13" Column="1" />
-    </Files>
-  </MonoDevelop.Ide.Workbench>
-  <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
-  <MonoDevelop.Ide.DebuggingService.Breakpoints>
-    <BreakpointStore />
-  </MonoDevelop.Ide.DebuggingService.Breakpoints>
-  <MonoDevelop.Ide.DebuggingService.PinnedWatches />
-  <MonoDevelop.Ide.ItemProperties.TrueProfile PreferredExecutionTarget="MonoDevelop.Default" />
-  <MultiItemStartupConfigurations />
-</Properties>
