--- C#/TrueProfile/bin/Debug/netcoreapp2.0/TrueProfile.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v2.0",
-    "signature": "da39a3ee5e6b4b0d3255bfef95601890afd80709"
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v2.0": {
-      "TrueProfile/1.0.0": {
-        "runtime": {
-          "TrueProfile.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "TrueProfile/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
