--- C#/TrueProfile/bin/Debug/netcoreapp2.0/TrueProfile.runtimeconfig.dev.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "/Users/divum/.dotnet/store/|arch|/|tfm|",
-      "/Users/divum/.nuget/packages",
-      "/usr/local/share/dotnet/sdk/NuGetFallbackFolder"
-    ]
-  }
-}
