--- C#/TrueProfile/bin/Debug/netcoreapp2.0/TrueProfile.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "netcoreapp2.0",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "2.0.0"
-    }
-  }
-}
